Check all shortest paths, not just one.
For distance to j:

For distance from each point i, If none of the neighbours of i that do not
share the cluster are part of the shortest path, then we do not leave
the cluster-sharing subgraph at i on any shortest path going
through i to j.

Since this is checked for every pair, we never leave the subgraph on any
shortest path between any two points in the subgraph.